# docs/swagger.yaml
basePath: /api
definitions:
...
  req.RequestUpdateProfile:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
...
  res.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: https://siphoria-api-production.up.railway.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HOTEL BOOKING API
  version: "1.0"
paths:
...
  /users/add-to-cart:
    post:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Add to cart
      tags:
      - user-service
  /users/carts:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Get cart by user
      tags:
      - user-service
  /users/details:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Get Customer Profile
      tags:
      - user-service
  /users/get-rank-available:
    get:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Get user rank
      tags:
      - user-service
  /users/notifications:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Get User Notifications
      tags:
      - user-service
  /users/update-avatar:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update user's avatar
      tags:
      - user-service
  /users/update-profile:
    patch:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update customer profile
      tags:
      - user-service
  /users/update-rank:
    post:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update user's rank
      tags:
      - user-service
swagger: "2.0"